+++ frontend/src/app/visualize/_components/ClassGroupNode.tsx
