+++ frontend/src/app/visualize/_components/DetailSidebar.tsx
