+++ frontend/src/app/visualize/_components/SVGMethodNode.tsx
